Update springboot3 and Java21

+++ src/main/java/com/disk91/etl/data/object/sub/WitnessHistory.java
@@ -14,7 +14,7 @@ public class WitnessHistory  implements ClonnableObject<WitnessHistory> {
     // ---------
 
     public void addTotLateMs(long deltaMs) {
-        this.totLateMs += deltaMs;
+        this.totLateMs += (int)deltaMs;
     }
 
     public WitnessHistory clone() {
